Extract verify annotation code
